requirements.txt
annotated-doc==0.0.4
antlr4-python3-runtime==4.9.3
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
einops==0.8.2
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
hydra-core==1.3.2
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
omegaconf==2.3.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
portalocker==3.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pywin32==311
PyYAML==6.0.3
regex==2026.1.15
safetensors==0.7.0
scipy==1.17.0
setuptools==69.5.1
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.3.0
timm==1.0.24
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer-slim==0.21.2
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.6.0
wheel==0.46.3
yacs==0.1.8
